+++ packages/waku/src/lib/builder.ts
@@ -20,8 +20,8 @@ import {
 } from './utils/node-fs.js';
 import { streamToString } from './utils/stream.js';
 import { encodeInput, generatePrefetchCode } from './rsc/utils.js';
-import { renderRSC, getBuildConfigRSC } from './rsc/renderer.js';
-import { renderHtml, shutdown as shutdownSsr } from './rsc/ssr.js';
+import { renderRsc, getBuildConfig } from './rsc/rsc-renderer.js';
+import { renderHtml, shutdown as shutdownSsr } from './rsc/html-renderer.js';
 import { rscIndexPlugin } from './vite-plugin/rsc-index-plugin.js';
 import { rscAnalyzePlugin } from './vite-plugin/rsc-analyze-plugin.js';
 import { rscTransformPlugin } from './vite-plugin/rsc-transform-plugin.js';
@@ -226,7 +226,7 @@ const buildClientBundle = async (
 };
 
 const emitRscFiles = async (config: ResolvedConfig) => {
-  const buildConfig = await getBuildConfigRSC({ config });
+  const buildConfig = await getBuildConfig({ config });
   const clientModuleMap = new Map<string, Set<string>>();
   const addClientModule = (input: string, id: string) => {
     let idSet = clientModuleMap.get(input);
@@ -257,7 +257,7 @@ const emitRscFiles = async (config: ResolvedConfig) => {
         if (!rscFileSet.has(destRscFile)) {
           rscFileSet.add(destRscFile);
           await mkdir(joinPath(destRscFile, '..'), { recursive: true });
-          const readable = await renderRSC({
+          const readable = await renderRsc({
             input,
             method: 'GET',
             config,
@@ -278,7 +278,7 @@ const emitRscFiles = async (config: ResolvedConfig) => {
 
 const emitHtmlFiles = async (
   config: ResolvedConfig,
-  buildConfig: Awaited<ReturnType<typeof getBuildConfigRSC>>,
+  buildConfig: Awaited<ReturnType<typeof getBuildConfig>>,
   getClientModules: (input: string) => string[],
   ssr: boolean,
 ) => {
