+++ packages/waku/src/main.ts
@@ -1,6 +1,6 @@
 export { honoMiddleware } from './lib/middleware/hono.js';
 export { connectMiddleware } from './lib/middleware/connect.js';
-export { createHandler } from './lib/rsc/handler.js';
+export { createHandler as unstable_createHandler } from './lib/rsc/handler.js';
 
 import type { build as buildOrig } from './lib/builder.js';
